content/events/test/test_bug607464.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=607464
-->
<head>
  <title>Test for Bug 607464</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=607464">Mozilla Bug 607464</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="application/javascript;version=1.7">

/**
 * Test for Bug 607464:
 * Pixel scrolling shouldn't scroll smoothly, even if general.smoothScroll is on.
 **/

function scrollDown15PxWithPixelScrolling(scrollbox) {
  var win = scrollbox.ownerDocument.defaultView;
  let event = {
    deltaMode: WheelEvent.DOM_DELTA_PIXEL,
    deltaY: 3.0,
    lineOrPageDeltaY: 1
  };
  // A pixel scroll with lineOrPageDeltaY.
  synthesizeWheel(scrollbox, 10, 10, event, win);
  // then 4 pixel scrolls without lineOrPageDeltaY.
  event.lineOrPageDeltaY = 0;
  for (let i = 0; i < 4; ++i) {
    synthesizeWheel(scrollbox, 10, 10, event, win);
  }

  // Note: the line scroll shouldn't have any effect because it has
  // hasPixels = true set on it. We send it to emulate normal
  // behavior.
}

function runTest() {
  var win = open('data:text/html,<!DOCTYPE html>\n' +
    '<div id="scrollbox" style="height: 100px; overflow: auto;">' +
    '  <div style="height: 1000px;"></div>' +
    '</div>', '_blank', 'width=300,height=300');
  SimpleTest.waitForFocus(function () {
    var scrollbox = win.document.getElementById("scrollbox");
    let scrollTopBefore = scrollbox.scrollTop;

    scrollDown15PxWithPixelScrolling(scrollbox);

    // wait for the next refresh driver run
    window.mozRequestAnimationFrame(function() {
      // actually, wait for the next one before checking results, since
      // scrolling might not be flushed until after this code has run
      window.mozRequestAnimationFrame(function() {
        is(scrollbox.scrollTop, scrollTopBefore + 15,
           "Pixel scrolling should have finished after one refresh driver iteration. " +
           "We shouldn't be scrolling smoothly, even though the pref is set.");
        win.close();
        clearPrefs();
        SimpleTest.finish();
      });
    });
  }, win);
}

function initPrefs()
{
  // Disables the app level scroll acceleration
  SpecialPowers.setIntPref("mousewheel.acceleration.start", -1);
  SpecialPowers.setBoolPref("mousewheel.system_scroll_override_on_root_content.enabled", false);

  // Enables smooth scrolling
  SpecialPowers.setBoolPref("general.smoothScroll", true);
}

function clearPrefs()
{
  SpecialPowers.clearUserPref("mousewheel.acceleration.start");
  SpecialPowers.clearUserPref("mousewheel.system_scroll_override_on_root_content.enabled");
  SpecialPowers.clearUserPref("general.smoothScroll");
}

window.onload = function () {
  initPrefs();
  SimpleTest.executeSoon(runTest);
}

SimpleTest.waitForExplicitFinish();

</script>
</pre>

</body>
</html>
